Refactor employees
