+++ src/main/java/uk/gov/companieshouse/api/accounts/Kind.java
@@ -10,14 +10,10 @@ public enum Kind {
     POLICY_NOTE("small-full-accounts-note#accounting-policy"),
     CREDITORS_AFTER_ONE_YEAR_NOTE("small-full-accounts-note#creditors-after-one-year"),
     CREDITORS_DUE_WITHIN_ONE_YEAR_NOTE("small-full-accounts-note#creditors-within-one-year"),
-    DEBTORS_NOTE("small-full-accounts-note#debtors"),
     INTANGIBLE_ASSETS_NOTE("small-full-accounts-note#intangible-assets"),
     SMALL_FULL_STATEMENT("small-full-accounts#statements"),
-    STOCKS_NOTE("small-full-accounts#stocks-note"),
-    TANGIBLE_ASSETS_NOTE("small-full-accounts-note#tangible-assets"),
     FIXED_ASSETS_INVESTMENTS_NOTE("small-full-accounts-note#fixed-assets-investments"),
     CURRENT_ASSETS_INVESTMENT_NOTE("small-full-accounts-note#current-assets-investments"),
-    EMPLOYEES_NOTE("small-full-accounts-note#employees"),
     CIC_REPORT("cic-report#cic-report"),
     CIC_APPROVAL("cic-report#approval"),
     CIC_STATEMENTS("cic-report#cic-statements"),
