--- src/main/java/uk/gov/companieshouse/api/accounts/service/impl/EmployeesService.java
@@ -1,185 +1,0 @@
-package uk.gov.companieshouse.api.accounts.service.impl;
-
-import com.mongodb.MongoException;
-import java.util.HashMap;
-import java.util.Map;
-import javax.servlet.http.HttpServletRequest;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DuplicateKeyException;
-import org.springframework.stereotype.Service;
-import uk.gov.companieshouse.GenerateEtagUtil;
-import uk.gov.companieshouse.api.accounts.Kind;
-import uk.gov.companieshouse.api.accounts.ResourceName;
-import uk.gov.companieshouse.api.accounts.exception.DataException;
-import uk.gov.companieshouse.api.accounts.links.BasicLinkType;
-import uk.gov.companieshouse.api.accounts.links.SmallFullLinkType;
-import uk.gov.companieshouse.api.accounts.model.entity.notes.employees.EmployeesEntity;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.employees.Employees;
-import uk.gov.companieshouse.api.accounts.model.validation.Errors;
-import uk.gov.companieshouse.api.accounts.repository.EmployeesRepository;
-import uk.gov.companieshouse.api.accounts.service.ResourceService;
-import uk.gov.companieshouse.api.accounts.service.response.ResponseObject;
-import uk.gov.companieshouse.api.accounts.service.response.ResponseStatus;
-import uk.gov.companieshouse.api.accounts.transformer.EmployeesTransformer;
-import uk.gov.companieshouse.api.accounts.utility.impl.KeyIdGenerator;
-import uk.gov.companieshouse.api.accounts.validation.EmployeesValidator;
-import uk.gov.companieshouse.api.model.transaction.Transaction;
-
-@Service
-public class EmployeesService implements ResourceService<Employees> {
-
-    private EmployeesRepository repository;
-    private EmployeesTransformer transformer;
-    private KeyIdGenerator keyIdGenerator;
-    private SmallFullService smallFullService;
-    private EmployeesValidator employeesValidator;
-
-    @Autowired
-    public EmployeesService(EmployeesRepository repository,
-            EmployeesTransformer transformer,
-            KeyIdGenerator keyIdGenerator,
-            SmallFullService smallFullService,
-            EmployeesValidator employeesValidator) {
-
-        this.repository = repository;
-        this.transformer = transformer;
-        this.keyIdGenerator = keyIdGenerator;
-        this.smallFullService = smallFullService;
-        this.employeesValidator = employeesValidator;
-    }
-
-    @Override
-    public ResponseObject<Employees> create(Employees rest,
-            Transaction transaction,
-            String companyAccountId,
-            HttpServletRequest request) throws DataException {
-
-        Errors errors = employeesValidator.validateEmployees(rest, transaction);
-
-        if (errors.hasErrors()) {
-
-            return new ResponseObject<>(ResponseStatus.VALIDATION_ERROR, errors);
-        }
-
-        setMetadataOnRestObject(rest, transaction, companyAccountId);
-
-        EmployeesEntity entity = transformer.transform(rest);
-        entity.setId(generateID(companyAccountId));
-
-        try {
-            repository.insert(entity);
-        } catch (DuplicateKeyException e) {
-
-            return new ResponseObject<>(ResponseStatus.DUPLICATE_KEY_ERROR);
-        } catch (MongoException e) {
-
-            throw new DataException(e);
-        }
-
-        smallFullService.addLink(companyAccountId, SmallFullLinkType.EMPLOYEES_NOTE,
-                getSelfLinkFromEmployeesEntity(entity), request);
-
-        return new ResponseObject<>(ResponseStatus.CREATED, rest);
-    }
-
-    @Override
-    public ResponseObject<Employees> update(Employees rest,
-            Transaction transaction,
-            String companyAccountId,
-            HttpServletRequest request) throws DataException {
-
-        Errors errors = employeesValidator.validateEmployees(rest, transaction);
-
-        if (errors.hasErrors()) {
-
-            return new ResponseObject<>(ResponseStatus.VALIDATION_ERROR, errors);
-        }
-
-        setMetadataOnRestObject(rest, transaction, companyAccountId);
-
-        EmployeesEntity entity = transformer.transform(rest);
-        entity.setId(generateID(companyAccountId));
-
-        try {
-            repository.save(entity);
-        } catch (MongoException me) {
-
-            throw new DataException(me);
-        }
-
-        return new ResponseObject<>(ResponseStatus.UPDATED, rest);
-    }
-
-    @Override
-    public ResponseObject<Employees> find(String companyAccountsId,
-            HttpServletRequest request) throws DataException {
-
-        EmployeesEntity entity;
-
-        try {
-            entity = repository.findById(generateID(companyAccountsId)).orElse(null);
-        } catch (MongoException e) {
-
-            throw new DataException(e);
-        }
-
-        if (entity == null) {
-            return new ResponseObject<>(ResponseStatus.NOT_FOUND);
-        }
-
-        return new ResponseObject<>(ResponseStatus.FOUND, transformer.transform(entity));
-    }
-
-    @Override
-    public ResponseObject<Employees> delete(String companyAccountsId,
-                                                             HttpServletRequest request) throws DataException {
-
-        String employeesId = generateID(companyAccountsId);
-
-        try {
-            if (repository.existsById(employeesId)) {
-                repository.deleteById(employeesId);
-
-                smallFullService.removeLink(companyAccountsId,
-                        SmallFullLinkType.EMPLOYEES_NOTE, request);
-                return new ResponseObject<>(ResponseStatus.UPDATED);
-            } else {
-                return new ResponseObject<>(ResponseStatus.NOT_FOUND);
-            }
-        } catch (MongoException me) {
-
-            throw new DataException(me);
-        }
-    }
-
-    private String generateID(String companyAccountId) {
-        return keyIdGenerator.generate(companyAccountId + "-" + ResourceName.EMPLOYEES.getName());
-    }
-
-    private void setMetadataOnRestObject(Employees rest, Transaction transaction,
-            String companyAccountsId) {
-
-        rest.setLinks(createSelfLink(transaction, companyAccountsId));
-        rest.setEtag(GenerateEtagUtil.generateEtag());
-        rest.setKind(Kind.EMPLOYEES_NOTE.getValue());
-    }
-
-    private Map<String, String> createSelfLink(Transaction transaction, String companyAccountsId) {
-
-        Map<String, String> map = new HashMap<>();
-        map.put(BasicLinkType.SELF.getLink(), generateSelfLink(transaction, companyAccountsId));
-        return map;
-    }
-
-    private String generateSelfLink(Transaction transaction, String companyAccountId) {
-
-        return transaction.getLinks().getSelf() + "/"
-                + ResourceName.COMPANY_ACCOUNT.getName() + "/" + companyAccountId + "/"
-                + ResourceName.SMALL_FULL.getName() + "/notes/" + ResourceName.EMPLOYEES.getName();
-    }
-
-
-    public String getSelfLinkFromEmployeesEntity(EmployeesEntity entity) {
-        return entity.getData().getLinks().get(BasicLinkType.SELF.getLink());
-    }
-}
